🐛 fix: seperate logic and ui in description components
